# Speedup retrieve all