--- a/setup.cfg
+++ b/setup.cfg
@@ -3,7 +3,7 @@ verbosity=2
 detailed-errors=1
 with-ignore-docstrings=1
 with-timer=1
-timer-top-n=15
+timer-filter=warning
 
 [metadata]
 description-file = README.rst