--- a/zipline/utils/control_flow.py
+++ b/zipline/utils/control_flow.py
@@ -1,6 +1,7 @@
 """
 Control flow utilities.
 """
+from six import iteritems
 from warnings import (
     catch_warnings,
     filterwarnings,
@@ -54,3 +55,19 @@ def ignore_nanwarnings():
             {'category': RuntimeWarning, 'module': 'numpy.lib.nanfunctions'},
         )
     )
+
+
+def invert(d):
+    """
+    Invert a dictionary into a dictionary of lists.
+
+    >>> invert({'a': 1, 'b': 2, 'c': 1})
+    {1: ['a', 'c'], 2: ['b']}
+    """
+    out = {}
+    for k, v in iteritems(d):
+        try:
+            out[v].append(k)
+        except KeyError:
+            out[v] = [k]
+    return out