… tests/pipeline/base.py
… tests/pipeline/test_blaze.py
… tests/pipeline/test_engine.py
… tests/pipeline/test_pipeline_algo.py
… tests/test_assets.py
… zipline/utils/test_utils.py