Speed up `AssetFinder.retrieve_all` by doing one query instead of N.

- Removes the separate caches for equities and futures in favor of a single asset cache.
- We now cache the results of failed lookups.  This required unfortunate hacks in the `TradingAlgorithm` code that resolves assets from DataFrames.
- Reimplements single-asset lookups as a special case of multiple-asset lookups.