@ssanderson 2015-11-05
Timings for ~30000 assets on my machine:

**Before:**
```
# Empty Cache
In [3]: %timeit -r1 -n1 f.retrieve_all(sids)
1 loops, best of 1: 29.4 s per loop

# Full Cache
In [5]: %timeit f.retrieve_all(sids)
100 loops, best of 3: 9.13 ms per loop

# Single Asset
In [3]: %timeit f.retrieve_asset(24)
The slowest run took 49.35 times longer than the fastest. This could mean that an intermediate result is being cached
1000000 loops, best of 3: 546 ns per loop
```

**After:**
```
# Empty Cache
In [5]: %timeit -n5 f.retrieve_all(sids); f._reset_caches()
5 loops, best of 3: 1.34 s per loop

# Full Cache
In [7]: %timeit f.retrieve_all(sids)
100 loops, best of 3: 5.96 ms per loop

# Single Asset
In [4]: %timeit f.retrieve_asset(24)
The slowest run took 42.91 times longer than the fastest. This could mean that an intermediate result is being cached
1000000 loops, best of 3: 278 ns per loop
```

**Conclusions:**
- This is roughly 30x faster for batch lookups on a large number of assets when caches are empty.
- This is roughly 2x faster for batch lookups when caches are full.
- This is roughly 2x **slower** for single-asset lookups.  Those lookups go from 270 ns to 540 ns on my machine.  I think this is okay because that performance should only matter appreciably if we're looking up lots of assets in a loop, in which case we should just use the faster batch mode anyway.  We can add more special-case handling for single-asset cache hits if we need to though.

cc @jbredeche on perf notes.