@richafrank 2015-11-05
Nice conclusions!  Is the single-asset lookup case for cache-miss or cache-hit?