@ssanderson 2015-11-06
@richafrank the single-asset case is about the same 2x slowdown for both hits and misses.