@ssanderson 2015-11-13
@mtydykov can you take a look at the commits from a436426a5a018605895d6d073ba709a5ee82cdef to 9881c303413534e8859a85d40c27aa5bffaf4a1a?  The rest of this PR has already been reviewed by @richafrank, but I made a couple changes after rebasing against your latest changes.